updating branch

+++ HospitalProject2/Controllers/VolunteersDataController.cs
@@ -1,4 +1,6 @@
-﻿using System;
+using System;
+using System.IO;
+using System.Web;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
@@ -11,6 +13,8 @@ using System.Web.Http.Description;
 using HospitalProject2.Models;
 using System.Diagnostics;
 
+using Microsoft.AspNet.Identity;
+
 
 namespace HospitalProject2.Controllers
 {
@@ -21,9 +25,22 @@ namespace HospitalProject2.Controllers
         // GET: api/VolunteersData/ListVolunteers
         [ResponseType(typeof(VolunteersDto))]
         [HttpGet]
+        [Authorize(Roles = "Admin,Guest")]
         public IHttpActionResult ListVolunteers()
         {
-            List<Volunteers> Volunteers = db.Volunteers.ToList();
+            bool isAdmin = User.IsInRole("Admin");
+            //Admins see all, guests only see their own
+            List<Volunteers> Volunteers;
+            Debug.WriteLine("id is " + User.Identity.GetUserId());
+            if (isAdmin) Volunteers = db.Volunteers.ToList();
+            else
+            {
+                string UserId = User.Identity.GetUserId();
+                Volunteers = db.Volunteers.Where(v => v.UserID == UserId).ToList();
+            }
+
+            //List<Volunteers> Volunteers = db.Volunteers.ToList();
+
             List<VolunteersDto> VolunteersDtos = new List<VolunteersDto>();
 
             Volunteers.ForEach(v => VolunteersDtos.Add(new VolunteersDto()
@@ -34,16 +51,40 @@ namespace HospitalProject2.Controllers
                 contact = v.contact,
                 email = v.email,
                 program_id = v.program_id,
-                hours = v.hours
+                hours = v.hours,
+                UserId = v.UserID,
+                name = v.Programs.name
 
             }));
       
             return Ok(VolunteersDtos);
         }
 
+        [HttpGet]
+        [ResponseType(typeof(VolunteersDto))]
+        public IHttpActionResult ListVolunteersForProgram(int id)
+        {
+            List<Volunteers> Volunteers = db.Volunteers.Where(v => v.program_id == id).ToList();
+            List<VolunteersDto> VolunteersDtos = new List<VolunteersDto>();
+
+            Volunteers.ForEach(v => VolunteersDtos.Add(new VolunteersDto()
+            {
+                volunteer_id = v.volunteer_id,
+                f_name = v.f_name,
+                l_name = v.l_name,
+                contact = v.contact,
+                email = v.email,
+                hours = v.hours,
+                UserId= v.UserID
+
+            }));
+            return Ok(VolunteersDtos);
+        }
+
         // GET: api/VolunteersData/FindVolunteers/5
         [ResponseType(typeof(Volunteers))]
         [HttpGet]
+        [Authorize(Roles = "Admin,Guest")]
         public IHttpActionResult FindVolunteers(int id)
         {
             Volunteers Volunteers = db.Volunteers.Find(id);
@@ -55,13 +96,20 @@ namespace HospitalProject2.Controllers
                 contact = Volunteers.contact,
                 email = Volunteers.email,
                 program_id = Volunteers.program_id,
-                hours = Volunteers.hours
+                hours = Volunteers.hours,
+                UserId = Volunteers.UserID,
+                name = Volunteers.Programs.name
+
             };
 
             if (Volunteers == null)
             {
                 return NotFound();
             }
+            //do not process if the (user is not an admin) and (the booking does not belong to the user)
+            bool isAdmin = User.IsInRole("Admin");
+            //Forbidden() isn't a natively implemented status like BadRequest()
+            if (!isAdmin && (Volunteers.UserID != User.Identity.GetUserId())) return StatusCode(HttpStatusCode.Forbidden);
 
             return Ok(VolunteersDto);
         }
@@ -69,6 +117,7 @@ namespace HospitalProject2.Controllers
         // POST: api/VolunteersData/UpdateVolunteers/5
         [ResponseType(typeof(void))]
         [HttpPost]
+        [Authorize(Roles = "Admin,Guest")]
         public IHttpActionResult UpdateVolunteers(int id, Volunteers Volunteers)
         {
             if (!ModelState.IsValid)
@@ -81,7 +130,17 @@ namespace HospitalProject2.Controllers
                 return BadRequest();
             }
 
+            //do not process if the (user is not an admin) and (the booking does not belong to the user)
+            bool isAdmin = User.IsInRole("Admin");
+            //Forbidden() isn't a natively implemented status like BadRequest()
+            if (!isAdmin && (Volunteers.UserID != User.Identity.GetUserId()))
+            {
+                Debug.WriteLine("not allowed. booking user" + Volunteers.UserID + " user " + User.Identity.GetUserId());
+                return StatusCode(HttpStatusCode.Forbidden);
+            }
+
             db.Entry(Volunteers).State = EntityState.Modified;
+            db.Entry(Volunteers).Property(v => v.UserID).IsModified = false;
 
             try
             {
@@ -105,12 +164,15 @@ namespace HospitalProject2.Controllers
         // POST: api/VolunteersData/AddVolunteers
         [ResponseType(typeof(Volunteers))]
         [HttpPost]
+        [Authorize(Roles = "Admin,Guest")]
         public IHttpActionResult AddVolunteers(Volunteers Volunteers)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
+
+            Volunteers.UserID = User.Identity.GetUserId();
 
             db.Volunteers.Add(Volunteers);
             db.SaveChanges();
@@ -121,6 +183,7 @@ namespace HospitalProject2.Controllers
         // POST: api/VolunteersData/DeleteVolunteers/5
         [ResponseType(typeof(Volunteers))]
         [HttpPost]
+        [Authorize(Roles = "Admin,Guest")]
         public IHttpActionResult DeleteVolunteers(int id)
         {
             Volunteers Volunteers = db.Volunteers.Find(id);
@@ -128,6 +191,12 @@ namespace HospitalProject2.Controllers
             {
                 return NotFound();
             }
+
+            //do not process if the (user is not an admin) and (the booking does not belong to the user)
+            bool isAdmin = User.IsInRole("Admin");
+            //Forbidden() isn't a natively implemented status like BadRequest()
+            if (!isAdmin && (Volunteers.UserID != User.Identity.GetUserId())) return StatusCode(HttpStatusCode.Forbidden);
+
 
             db.Volunteers.Remove(Volunteers);
             db.SaveChanges();
